allow users to edit the platform path if it wasn't verified
we don't allow users to change the platform path if a makefile was
specified, however, because the verify may have ran and fail because
of a faulty makefile in which case some cruft from the drush make run
may still be lying around in the backend. we don't have this problem
if there's no makefile because it's the admin's responsability to
cleanup after themselves then.

Closes: #1898260